refactor(fees): Move fees class methods to fetcher class
