+++ src/index.ts
@@ -15,8 +15,6 @@ export {
   parseBigintIsh
 } from './utils'
 
-export * from './fees'
-
 export * from './errors'
 export * from './entities'
 export * from './router'
