+++ test/fees.test.ts
@@ -1,4 +1,4 @@
-import { ChainId, Pair, WETH, DXD, Fees, TEST_TOKENS, JSBI } from '../src'
+import { ChainId, Pair, WETH, DXD, Fetcher, TEST_TOKENS, JSBI } from '../src'
 import { ZERO_ADDRESS } from '../src/constants'
 
 describe('fees', () => {
@@ -6,9 +6,9 @@ describe('fees', () => {
     it('Get WETH-DXD kovan fee', async () => {
       const WETH_DXD_KOVAN = await Pair.fetchData(WETH[ChainId.KOVAN], DXD[ChainId.KOVAN])
       const DXD_WEENUS_KOVAN = await Pair.fetchData(DXD[ChainId.KOVAN], TEST_TOKENS.WEENUS[ChainId.KOVAN])
-      const DXDKovanFee = await Fees.fetchSwapFee(WETH_DXD_KOVAN.liquidityToken)
+      const DXDKovanFee = await Fetcher.fetchSwapFee(WETH_DXD_KOVAN.liquidityToken)
       expect(DXDKovanFee.fee).toEqual(JSBI.BigInt(20))
-      const DXD_WEENUS_KOVAN_FEES = await Fees.fetchSwapFee(DXD_WEENUS_KOVAN.liquidityToken)
+      const DXD_WEENUS_KOVAN_FEES = await Fetcher.fetchSwapFee(DXD_WEENUS_KOVAN.liquidityToken)
       expect(DXD_WEENUS_KOVAN_FEES.fee).toEqual(JSBI.BigInt(30))
     })
   })
@@ -17,7 +17,7 @@ describe('fees', () => {
     it('Get WETH-DXD kovan fee', async () => {
       const WETH_DXD_KOVAN = await Pair.fetchData(WETH[ChainId.KOVAN], DXD[ChainId.KOVAN])
       const DXD_WEENUS_KOVAN = await Pair.fetchData(DXD[ChainId.KOVAN], TEST_TOKENS.WEENUS[ChainId.KOVAN])
-      const fees = await Fees.fetchSwapFees([WETH_DXD_KOVAN.liquidityToken, DXD_WEENUS_KOVAN.liquidityToken])
+      const fees = await Fetcher.fetchSwapFees([WETH_DXD_KOVAN.liquidityToken, DXD_WEENUS_KOVAN.liquidityToken])
       expect(fees[0].fee).toEqual(JSBI.BigInt(20))
       expect(fees[1].fee).toEqual(JSBI.BigInt(30))
     })
@@ -28,7 +28,7 @@ describe('fees', () => {
       const WETH_DXD_KOVAN = await Pair.fetchData(WETH[ChainId.KOVAN], DXD[ChainId.KOVAN])
       const WETH_WEENUS_KOVAN = await Pair.fetchData(WETH[ChainId.KOVAN], TEST_TOKENS.WEENUS[ChainId.KOVAN])
       const DXD_WEENUS_KOVAN = await Pair.fetchData(DXD[ChainId.KOVAN], TEST_TOKENS.WEENUS[ChainId.KOVAN])
-      const fees = await Fees.fetchAllSwapFees(ChainId.KOVAN)
+      const fees = await Fetcher.fetchAllSwapFees(ChainId.KOVAN)
       expect(fees[WETH_DXD_KOVAN.liquidityToken.address].fee).toEqual(JSBI.BigInt(20))
       expect(fees[WETH_WEENUS_KOVAN.liquidityToken.address].fee).toEqual(JSBI.BigInt(30))
       expect(fees[DXD_WEENUS_KOVAN.liquidityToken.address].fee).toEqual(JSBI.BigInt(30))
@@ -38,7 +38,7 @@ describe('fees', () => {
       const WETH_DXD_KOVAN = await Pair.fetchData(WETH[ChainId.KOVAN], DXD[ChainId.KOVAN])
       const WETH_WEENUS_KOVAN = await Pair.fetchData(WETH[ChainId.KOVAN], TEST_TOKENS.WEENUS[ChainId.KOVAN])
       const DXD_WEENUS_KOVAN = await Pair.fetchData(DXD[ChainId.KOVAN], TEST_TOKENS.WEENUS[ChainId.KOVAN])
-      const fees = await Fees.fetchAllSwapFees(ChainId.KOVAN, {
+      const fees = await Fetcher.fetchAllSwapFees(ChainId.KOVAN, {
         [WETH_DXD_KOVAN.liquidityToken.address] : {
           fee: JSBI.BigInt(20),
           owner: ZERO_ADDRESS
@@ -56,7 +56,7 @@ describe('fees', () => {
   
   describe('fetchProtocolFee', () => {
     it('Get WETH-DXD kovan fee', async () => {
-      const protocolFee = await Fees.fetchProtocolFee(ChainId.KOVAN)
+      const protocolFee = await Fetcher.fetchProtocolFee(ChainId.KOVAN)
       expect(protocolFee.feeDenominator).toEqual(5)
       expect(protocolFee.feeReceiver).toEqual(ZERO_ADDRESS)
     })
